Allow empty body; refactors DATA reader section - early errors
Empty smtp message body results in an empty buffer that produces
io.EOF. We now make an empty buffer have a single line so this
error does not happen.

The code handled errors at the end of conditionals, in the `else`,
rather than handling errors immediately and continuing/returning early.
Handling errors and stopping early results in easier to read code
and less if/else nesting.